break out inputs view
