+++ src/pages/scientificServices/pipelines/tabs/history/details/sections/io/JobOutputsView.tsx
